Created `LinguisticContext`.
