+++ docs/_static/example.py
@@ -1,8 +1,7 @@
 from nlglib.aggregation import aggregate
 from nlglib import pipeline
-from nlglib import nlg
 from nlglib.lexicon import NP, NNP, VP, AdjP, Tense, Aspect, Number, Noun, Gender
-from nlglib.realisation import Realiser
+from nlglib.realisation.backends.simplenlg import realise
 from nlglib.structures import String, Clause, Coordination, Var
 
 import logging
@@ -11,15 +10,14 @@ logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.ERROR)
 
 
 def run_simple_examples():
-    realiser = Realiser()
     s = String('This is my string')
-    print(realiser.realise(s))
+    print(realise(s))
 
     s = Clause(NNP('John'), VP('be', AdjP('happy')))
-    print(realiser.realise(s))
+    print(realise(s))
 
     s = Clause(NNP('Paul'), VP('play', NP('guitar'), features=dict([Aspect.progressive])))
-    print(realiser.realise(s))
+    print(realise(s))
 
     guitarists = Coordination(Clause(NNP('John'),
                                      VP('play', NP('a', 'guitar'),
@@ -31,8 +29,8 @@ def run_simple_examples():
                                      VP('play', NP('a', 'guitar'),
                                         features=dict([Aspect.progressive, Tense.past])))
                               )
-    print(realiser.realise(guitarists))
-    print(realiser.realise(aggregate(guitarists)))
+    print(realise(guitarists))
+    print(realise(aggregate(guitarists)))
 
     gringo = Coordination(Clause(NNP('George'),
                                  VP('play', NP('a', 'base'),
@@ -41,8 +39,8 @@ def run_simple_examples():
                                  VP('play', NP('drum', features=dict([Number.plural])),
                                     features=dict([Aspect.progressive, Tense.past])))
                           )
-    print(realiser.realise(gringo))
-    print(realiser.realise(aggregate(gringo)))
+    print(realise(gringo))
+    print(realise(aggregate(gringo)))
 
 
 def run_pipeline():
@@ -57,16 +55,16 @@ def run_pipeline():
         'Happy': Clause(NP(Var(0)), VP('be', AdjP('happy'))),
         'Play': Clause(NP(Var(0)), VP('play', NP(Var(1)))),
     }
-    input_str = 'Play(john, guitar) & Play(paul, guitar); Play(george, bass); Play(ringo, drums)'
-    output_str = pipeline.translate(input_str, templates, [])
+    input_str = 'Play(john, guitar) & Play(paul, guitar); ' \
+                'Play(george, bass); Play(ringo, drums)'
+    p = pipeline.Pipeline(__name__)
+    output_str = p.process(input_str, templates=templates)
     print(output_str)
 
 
 if __name__ == '__main__':
-    nlg.init_from_settings()
     print('starting')
     run_simple_examples()
     print('*' * 80)
     run_pipeline()
     print('done')
-    nlg.shutdown()
